Moved FromBridgedChainMessagesProof to bp-messages (#2170)
* moved FromBridgedChainMessagesProof to bp-messages

* fmt
