+++ bridges/bin/runtime-common/src/mock.rs
@@ -23,7 +23,7 @@ use crate::messages::{
 		FromThisChainMaximalOutboundPayloadSize, FromThisChainMessagePayload,
 		TargetHeaderChainAdapter,
 	},
-	target::{FromBridgedChainMessagePayload, SourceHeaderChainAdapter},
+	target::SourceHeaderChainAdapter,
 	BridgedChainWithMessages, HashOf, MessageBridge, ThisChainWithMessages,
 };
 
@@ -209,7 +209,7 @@ impl pallet_bridge_messages::Config for TestRuntime {
 	type MaximalOutboundPayloadSize = FromThisChainMaximalOutboundPayloadSize<OnThisChainBridge>;
 	type OutboundPayload = FromThisChainMessagePayload;
 
-	type InboundPayload = FromBridgedChainMessagePayload;
+	type InboundPayload = Vec<u8>;
 	type InboundRelayer = BridgedChainAccountId;
 	type DeliveryPayments = ();
